before Modularization

+++ botMerval/src/main/java/developBot/MervalOperations/models/clientModels/miCuenta/portafolio/Posicion.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class Posicion {
-    private Long cantidad;
+    private Integer cantidad;
     private Double comprometido;
     private Double puntosVariacion;
     private Double variacionDiaria;
